Support for iterables, ranges, function declarations and for loops

+++ src/main/kotlin/io/github/cafeteriaguild/lin/rt/lib/LNumber.kt
@@ -33,7 +33,7 @@ interface LNumber : LObj {
             return when {
                 left is Double || right is Double -> left.toDouble() + right.toDouble()
                 left is Float || right is Float -> left.toFloat() + right.toFloat()
-                left is Long || right is Long -> left.toLong() + right.toFloat()
+                left is Long || right is Long -> left.toLong() + right.toLong()
                 else -> left.toInt() + right.toInt()
             }
         }
